Fix Footpaths being dominated by worse Trips
In some cases a FootConnection would have been dominated by a
TripConnection that has less connecting trains. This is due to the fact
that FootConnections don't consider the change time at the target stop.

We work around this by rating a connection also by it's earliest reachable
time. For the target stop we can provide a much stricter Pareto set which
does not use criteria that is only relevant for traveling further.